optimzed rest interface and switched to quarkus on client

+++ server/src/test/java/at/meks/backupclientserver/context/backup/adapter/rest/BackupWebServiceTest.java
@@ -2,7 +2,6 @@ package at.meks.backupclientserver.context.backup.adapter.rest;
 
 import at.meks.backupclientserver.context.infrastructure.Configuration;
 import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.apache.commons.io.FileUtils;
 import org.apache.http.HttpStatus;
@@ -10,22 +9,23 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import javax.inject.Inject;
-
 import java.io.IOException;
 import java.nio.file.Files;
 
 import static io.restassured.RestAssured.given;
-import static org.junit.jupiter.api.Assertions.*;
 
 @QuarkusTest
 class BackupWebServiceTest {
 
+    public static final String BACKUP_URL = "/api/v1.0/{clientId}/{directory}/{filename}";
     @Inject
     Configuration configuration;
 
     @BeforeEach
     public void deleteUploadFolder() throws IOException {
-        FileUtils.forceDelete(configuration.uploadDir().toFile());
+        if (Files.exists(configuration.uploadDir())) {
+            FileUtils.forceDelete(configuration.uploadDir().toFile());
+        }
     }
 
     @Test
@@ -34,7 +34,7 @@ class BackupWebServiceTest {
                 .contentType(ContentType.BINARY)
                 .body(getClass().getResourceAsStream("fileVersion1.txt"))
                 .when()
-                .post("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .put(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
@@ -43,14 +43,14 @@ class BackupWebServiceTest {
                 .contentType(ContentType.BINARY)
                 .body(getClass().getResourceAsStream("fileVersion2.txt"))
                 .when()
-                .put("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .put(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
 
         given()
                 .when()
-                .delete("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .delete(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
@@ -60,7 +60,7 @@ class BackupWebServiceTest {
                 .contentType(ContentType.BINARY)
                 .body(getClass().getResourceAsStream("fileVersion2.txt"))
                 .when()
-                .put("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .put(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
@@ -72,7 +72,7 @@ class BackupWebServiceTest {
                 .contentType(ContentType.BINARY)
                 .body(getClass().getResourceAsStream("fileVersion1.txt"))
                 .when()
-                .put("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .put(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
@@ -80,11 +80,10 @@ class BackupWebServiceTest {
                 .contentType(ContentType.BINARY)
                 .body(getClass().getResourceAsStream("fileVersion1.txt"))
                 .when()
-                .put("/api/v1.0/backup/{clientId}/{directory}/{filename}",
+                .put(BACKUP_URL,
                         "testclient", "/home/user1/documents", "testFile.txt")
                 .then()
                 .statusCode(HttpStatus.SC_BAD_REQUEST);
     }
 
 }
-
